Removed MessageSection in favor of TransactionMessage
