+++ src/TransactionMessage.php
@@ -4,17 +4,18 @@ declare(strict_types=1);
 
 namespace EdifactParser;
 
-use EdifactParser\ReadModel\MessageSection;
 use EdifactParser\Segments\SegmentInterface;
 use EdifactParser\Segments\UNHMessageHeader;
 
 /** @psalm-immutable */
 final class TransactionMessage
 {
+    /** @var array<string, array<string,SegmentInterface>> */
+    private array $groupedSegments;
+
     /**
      * @psalm-pure
-     *
-     * @return MessageSection[]
+     * @psalm-return list<TransactionMessage>
      */
     public static function fromSegmentedValues(SegmentInterface...$segments): array
     {
@@ -24,15 +25,39 @@ final class TransactionMessage
         foreach ($segments as $segment) {
             if ($segment instanceof UNHMessageHeader) {
                 if ($segmentsGroup) {
-                    $messages[] = MessageSection::fromSegments(...$segmentsGroup);
+                    $messages[] = self::fromSegments(...$segmentsGroup);
                 }
                 $segmentsGroup = [];
             }
             $segmentsGroup[] = $segment;
         }
 
-        $messages[] = MessageSection::fromSegments(...$segmentsGroup);
+        $messages[] = self::fromSegments(...$segmentsGroup);
 
         return $messages;
+    }
+
+    /** @psalm-pure */
+    public static function fromSegments(SegmentInterface...$segments): self
+    {
+        $return = [];
+
+        foreach ($segments as $s) {
+            $return[$s->name()] ??= [];
+            $return[$s->name()][$s->subSegmentKey()] = $s;
+        }
+
+        return new self($return);
+    }
+
+    /** @param array<string, array<string,SegmentInterface>> $groupedSegments */
+    public function __construct(array $groupedSegments)
+    {
+        $this->groupedSegments = $groupedSegments;
+    }
+
+    public function segmentByName(string $name): array
+    {
+        return $this->groupedSegments[$name] ?? [];
     }
 }
