+++ tests/Functional/EdifactParserTest.php
@@ -6,7 +6,7 @@ namespace EdifactParser\Tests\Functional;
 
 use EdifactParser\EdifactParser;
 use EdifactParser\Exception\InvalidFile;
-use EdifactParser\ReadModel\MessageSection;
+
 use EdifactParser\Segments\CNTControl;
 use EdifactParser\Segments\SegmentInterface;
 use EdifactParser\Segments\UNHMessageHeader;
@@ -56,23 +56,22 @@ EDI;
 
         $transactionResult = EdifactParser::create()->parse($fileContent);
         self::assertCount(1, $transactionResult);
-        /** @var MessageSection $section */
-        $section = $transactionResult[0];
+        $message = $transactionResult[0];
 
         /** @var UNHMessageHeader $unh */
-        $unh = $section->segmentByName(UNHMessageHeader::class)['1'];
+        $unh = $message->segmentByName(UNHMessageHeader::class)['1'];
         self::assertEquals(['UNH', '1', ['IFTMIN', 'S', '93A', 'UN', 'PN001']], $unh->rawValues());
 
         /** @var CNTControl $cnt7 */
-        $cnt7 = $section->segmentByName(CNTControl::class)['7'];
+        $cnt7 = $message->segmentByName(CNTControl::class)['7'];
         self::assertEquals(['CNT', ['7', '0.1', 'KGM']], $cnt7->rawValues());
 
         /** @var CNTControl $cnt11 */
-        $cnt11 = $section->segmentByName(CNTControl::class)['11'];
+        $cnt11 = $message->segmentByName(CNTControl::class)['11'];
         self::assertEquals(['CNT', ['11', '1', 'PCE']], $cnt11->rawValues());
 
         /** @var UNTMessageFooter $unt */
-        $unt = $section->segmentByName(UNTMessageFooter::class)['19'];
+        $unt = $message->segmentByName(UNTMessageFooter::class)['19'];
         self::assertEquals(['UNT', '19', '1'], $unt->rawValues());
     }
 
@@ -90,14 +89,14 @@ EDI;
         $parser = EdifactParser::create(new TestingSegmentFactory('CUSTOM'));
         $transactionResult = $parser->parse($fileContent);
         self::assertCount(1, $transactionResult);
-        $section = $transactionResult[0];
+        $message = $transactionResult[0];
 
         /** @var SegmentInterface $custom */
-        $custom = $section->segmentByName('CUSTOM')['anyKey'];
+        $custom = $message->segmentByName('CUSTOM')['anyKey'];
         self::assertEquals(['CUSTOM', 'anyKey', ['whatever', 'value', '9']], $custom->rawValues());
 
         /** @var CNTControl $cnt11 */
-        $cnt11 = $section->segmentByName(CNTControl::class)['11'];
+        $cnt11 = $message->segmentByName(CNTControl::class)['11'];
         self::assertEquals(['CNT', ['11', '1', 'PCE']], $cnt11->rawValues());
     }
 }
